Move legend function to legend.py
